+++ tests/test_basic.py
@@ -108,27 +108,43 @@ def test_packages_create_delete(client, dataset):
     #assert pkg not in dataset
 
 def test_package_type_count(client,dataset):
+    n = dataset.package_count()
     pkg = DataPackage('Some MRI', package_type='MRI')
     assert not pkg.exists
     # create
     dataset.add(pkg)
     assert pkg.exists
+    client.update(pkg)
 
     pkg = DataPackage('Something else', package_type='TimeSeries')
     assert not pkg.exists
     dataset.add(pkg)
     assert pkg.exists
+    client.update(pkg)
 
-    n = dataset.get_package_type_count()
-    assert n == 2
+    m = dataset.package_count()
+    assert m == n + 2
 
 def test_publish_info(client,dataset):
-    publish_info = dataset.get_publish_info()
+    print(dataset.items)
+    publish_info = dataset.publish_info()
     assert publish_info.status =='NOT_PUBLISHED'
-    assert publish_info.latest_doi == None
-    assert publish_info.publishedVersionCount == 0
-    assert publish_info.lastPublishedDate == None
-    assert publish_info.latest_doi == None
+    assert publish_info.published_version_count == 0
+    assert publish_info.last_published== None
+    assert publish_info.doi == None
+
+def test_owner(client,dataset):
+    owner = dataset.owner()
+    assert owner.email == client.profile.email
+
+def test_collaborator_user(client,dataset):
+    collaborators = dataset.collaborator_users()
+    assert len(collaborators) == 1
+    assert collaborators[0].email == client.profile.email
+
+def test_collaborator_team(client,dataset):
+    collaborators = dataset.collaborator_teams()
+    assert len(collaborators) == 0
 
 def test_properties(client, dataset):